Add files via upload
Jane Street Market Prediction Kaggle Competition is a challenging quantitative trading competition sponsored by Jane Street. Participants are tasked with building a model that can quickly analyze short-term trade opportunities and determine whether to execute or dismiss them based on a set of anonymous 130 features.

The competition proves to be challenging owing to the enigmatic nature of certain features, the ever-evolving associations between features and responses, the uncertainty stemming from volatility, the low signal-to-noise ratio, and lastly, the fact that the trading nature of the data may cause it to behave differently in a testing set.

Additionally, the model must operate in real-time, with an average iteration time constraint of roughly 16 ms. As opportunities come up one at a time in a loop, the inference is done within this time constraint