Add files via upload
In this project, we will use the deduplication and blocking technique in order to reduce the number of comparisons for large dataset. Data deduplication can be applied in numerous fields of computing in research and other applications, such computer vision, NLP etc, but also importantly is reffered to a single instance storage, avoiding repeated data in the storage system.

The objective of this exercise is to deduplicate records from a dataset containing information about people.